eliminate async import
